perf: dedupe GetSimpleTypeName into shared TypeNameFormatter
Extract the duplicated runtime GetSimpleTypeName logic from
TestDataFormatter and DisplayNameBuilder into a single internal
TypeNameFormatter helper in TUnit.Core.Helpers.

The shared implementation builds the entire recursive name with one
pooled StringBuilder, allocating only the final string instead of a
new string[] + string.Join per generic level. Output is byte-identical
to both previous call sites.

Closes #6039
